Implement generic file blob fetching function and refactor URLPreviewComponent and Chat component
- Added a new `getFileBlob` function for centralized file retrieval from Azure blob storage.
- Refactored `URLPreviewComponent` to utilize the new `getFileBlob` function, removing redundant code.
- Updated the `Chat` component to use the centralized blob fetching logic, improving maintainability and consistency.
- Enhanced error handling and logging for file retrieval operations.
